+++ R/global_variables.R
@@ -151,7 +151,7 @@ if (getRversion() >= "2.15.1") {
       "filter.individuals.coverage.iqr",
       "filter.individuals.coverage.median", "COMMON_META", "where", "everything",
       "M_SEQ", "ID_SEQ", "1", "2", "ORIG_GT_BIN", "STRATA_SEQ", "IDTEMP", ".data",
-      "EQUAL_COUNTS", "strataG", "HET_EXP"
+      "EQUAL_COUNTS", "strataG", "HET_EXP", "id", "stratum"
     )
   )
 }